make maximum shields an int again

+++ server/src/main/kotlin/net/horizonsend/ion/server/features/starship/Starship.kt
@@ -485,8 +485,8 @@ class Starship(
 		get() = (shields.size.d().pow(0.9) / (initialBlockCount / 500.0).coerceAtLeast(1.0).pow(0.7))
 			.coerceAtMost(1.0)
 
-	val maxShields: Double = (0.00671215 * initialBlockCount.toDouble().pow(0.836512) - 0.188437)
-		get() = if (initialBlockCount < 500) field.coerceAtLeast(1.0) else field
+	val maxShields: Int = (0.00671215 * initialBlockCount.toDouble().pow(0.836512) - 0.188437).toInt()
+		get() = if (initialBlockCount < 500) field.coerceAtLeast(1) else field
 
 	val thrusterMap = mutableMapOf<BlockFace, ThrustData>()
